+++ Client/src/main/java/de/turtle_exception/core/client/api/TurtleClient.java
@@ -3,6 +3,7 @@ package de.turtle_exception.core.client.api;
 import de.turtle_exception.core.client.api.entities.Group;
 import de.turtle_exception.core.client.api.entities.User;
 import de.turtle_exception.core.client.api.entities.attribute.IUserContainer;
+import de.turtle_exception.core.client.api.requests.Action;
 import org.jetbrains.annotations.NotNull;
 
 import java.util.List;
